Refactor package for variation attribute edit classes
